zmiany w grafice
