+++ index.d.ts
@@ -1,0 +1,3 @@
+declare module '*.mp3'
+declare module '*.wav'
+declare module '*.ttf'
