+++ package.json
@@ -21,6 +21,11 @@
     "three": "0.132.2"
   },
   "devDependencies": {
-    "@types/three": "0.131.0"
+    "@types/css-font-loading-module": "0.0.6",
+    "@types/node": "^16.7.13",
+    "@types/three": "0.131.0",
+    "cssnano": "^5.0.8",
+    "postcss": "^8.2.7",
+    "typescript": "^4.4.2"
   }
 }
