+++ src/index.html
@@ -17,19 +17,11 @@
         padding: 0px;
         overflow: hidden;
       }
-      #canvas2D {
-        background-color: hsl(50, 100, 50);
-      }
-      #canvas3D {
-        background-color: hsl(50, 100, 50);
-        mix-blend-mode: lighten;
-      }
     </style>
   </head>
   <body>
-    <canvas id="canvas2D"></canvas>
-    <canvas id="canvas3D"></canvas>
+    <canvas  id="canvas2D"></canvas>
+    <canvas  id="canvas3D"></canvas>
   </body>
-
-  <script type="module" src="js/game.js"></script>
+    <script src="js/game.ts"></script>
 </html>
